Move bash command out of commands dir
The bash-command has been moved as it itself is not a command.
Instead it is used to run `power` and `console`. This is required
for the autocomplete to work correctly
