+++ src/bash.rb
