+++ src/commands.rb
@@ -6,4 +6,4 @@ require 'commands/hosts'
 require 'commands/hunter'
 require 'commands/dhcp'
 require 'commands/build'
-require 'commands/bash'
+require 'bash'
